chore(theme): separation of concerns
